Release Bug Fixes

+++ Services/StudyCourseService/StudyCourseService.cs
@@ -1498,6 +1498,7 @@ namespace griffined_api.Services.StudyCourseService
             return response;
         }
 
+
         public async Task<ServiceResponse<string>> EaRemoveStudent(EaStudentManagementRequestDto requestDto)
         {
             var response = new ServiceResponse<string>();
@@ -1535,10 +1536,12 @@ namespace griffined_api.Services.StudyCourseService
                 foreach (var dbStudyClass in dbStudySubject.StudyClasses)
                 {
                     var attendanceToRemove = dbStudyClass.Attendances
-                                            .FirstOrDefault(a => a.Student!.Id == requestDto.StudentId)
-                                            ?? throw new NotFoundException($"No Attendance with Student {requestDto.StudentId} Found.");
+                                            .FirstOrDefault(a => a.Student!.Id == requestDto.StudentId);
 
-                    _context.StudentAttendances.Remove(attendanceToRemove);
+                    if (attendanceToRemove is not null)
+                    {
+                        _context.StudentAttendances.Remove(attendanceToRemove);
+                    }
                 }
 
                 _context.StudySubjectMember.Remove(studentToRemove);
@@ -1677,6 +1680,7 @@ namespace griffined_api.Services.StudyCourseService
                                 .ThenInclude(sc => sc.Course)
                             .Include(s => s.Subject)
                             .Include(s => s.StudyClasses)
+                                .ThenInclude(x => x.Schedule)
                             .Include(s => s.StudyCourse)
                             .Include(s => s.StudySubjectMember)
                                 .ThenInclude(s => s.Student)
@@ -1697,7 +1701,6 @@ namespace griffined_api.Services.StudyCourseService
                                 ?? throw new NotFoundException($"Teacher ID {newSchedule.TeacherId} is not found.");
 
                     int studyClassCount = updateRequest.NewSchedule.Where(x => x.StudySubjectId == dbStudySubject.Id).Count();
-                    int c = 0;
 
                     var studyClass = new StudyClass
                     {
@@ -1713,17 +1716,6 @@ namespace griffined_api.Services.StudyCourseService
                         },
                     };
 
-                    if (c == studyClassCount / 2)
-                    {
-                        studyClass.IsFiftyPercent = true;
-                        studyClass.IsHundredPercent = false;
-                    }
-
-                    if (c == studyClassCount)
-                    {
-                        studyClass.IsFiftyPercent = false;
-                        studyClass.IsHundredPercent = true;
-                    }
 
                     var worktypes = _teacherService.GetTeacherWorkTypesWithHours(dbTeacher, newSchedule.Date.ToDateTime(), newSchedule.FromTime.ToTimeSpan(), newSchedule.ToTime.ToTimeSpan());
                     foreach (var worktype in worktypes)
@@ -1781,6 +1773,51 @@ namespace griffined_api.Services.StudyCourseService
             }
 
             await _context.SaveChangesAsync();
+
+            foreach (var studySubject in dbStudySubjects)
+            {
+                int c = 0;
+
+                var studyClasses = studySubject.StudyClasses
+                                               .OrderBy(sc => sc.Schedule.Date)
+                                               .ThenBy(sc => sc.Schedule.FromTime)
+                                               .Where(x => x.Status != ClassStatus.DELETED && x.Status != ClassStatus.CANCELLED)
+                                               .ToList();
+
+                var studyClassCount = studyClasses.Count;
+
+                foreach (var studyClass in studyClasses)
+                {
+                    c++;
+                    studyClass.IsFiftyPercent = false;
+                    studyClass.IsHundredPercent = false;
+                    if (c == studyClassCount / 2)
+                    {
+                        studyClass.IsFiftyPercent = true;
+                        studyClass.IsHundredPercent = false;
+                    }
+
+                    if (c == studyClassCount)
+                    {
+                        studyClass.IsFiftyPercent = false;
+                        studyClass.IsHundredPercent = true;
+                    }
+                }
+
+                for (int i = 0; i < studyClasses.Count; i++)
+                {
+                    studyClasses[i].ClassNumber = i + 1;
+                }
+
+                foreach (var studyClass in studySubject.StudyClasses.Where(x => x.Status == ClassStatus.DELETED || x.Status == ClassStatus.CANCELLED))
+                {
+                    studyClass.IsFiftyPercent = false;
+                    studyClass.IsHundredPercent = false;
+                }
+            }
+
+            await _context.SaveChangesAsync();
+
 
             var response = new ServiceResponse<string>
             {
